+++ gst-libs/gst/droid/Makefile.am
@@ -1,3 +1,4 @@
+AM_CPPFLAGS = -DSYSCONFDIR=\"$(sysconfdir)\"
 lib_LTLIBRARIES = libgstdroid-@GST_API_VERSION@.la
 libgstdroid_@GST_API_VERSION@_ladir = $(libdir)
 
@@ -15,10 +16,12 @@ libgstdroid_@GST_API_VERSION@_la_SOURCES = \
 	gstdroidmediabuffer.c \
 	gstdroidbufferpool.c \
 	gstdroidquery.c \
+	gstdroidcodec.c \
 	/usr/share/droidmedia/hybris.c
 
 libgstdroid_@GST_API_VERSION@_la_include_HEADERS = \
 	gstwrappedmemory.h \
 	gstdroidmediabuffer.h \
 	gstdroidbufferpool.h \
-	gstdroidquery.h
+	gstdroidquery.h \
+	gstdroidcodec.h
