+++ gst-libs/gst/droid/gstdroidcodec.h
