+++ gst/droidcodec/Makefile.am
@@ -10,13 +10,11 @@ libgstdroidcodec_la_LIBADD = $(GST_LIBS) $(NGI_LIBS) $(EGL_LIBS) $(DROID_LIBS) $
 libgstdroidcodec_la_SOURCES = \
 	gstdroidvdec.c \
 	gstdroidvenc.c \
-	gstdroidcodec.c \
 	gstdroidadec.c \
 	gstdroidaenc.c
 
 noinst_HEADERS = \
 	gstdroidvdec.h \
 	gstdroidvenc.h \
-	gstdroidcodec.h \
 	gstdroidadec.h \
 	gstdroidaenc.h
